requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证和授权
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 环境配置
python-dotenv==1.0.0

# 测试相关
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 日志和监控
structlog==23.2.0

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0 